backEnd/requirements.txt
fastapi
uvicorn
numpy
pandas
sqlalchemy
pydantic
pydantic-settings
python-multipart
python-dotenv
pyodbc
